meshlink_open: add log params + priv pointer
To log the setup, the log callback and private pointer needs to be set ahead, but it can only be set for an existing meshlink instance so it needs to be set by the open call.
Alternatively the open call needs to be split to first create a meshlink instance, then allow set the logging and priv, and only thereafter by another call open and setup the sockets.