Move the routing header out of the SPTPS payload.
Give every node in the mesh a unique ID (based on a hash of its name), and
add the ID of the source and destination packet at the start of a UDP
packet, before the SPTPS payload. This allows fast forwarding of packets,
and avoids having to look up nodes based on UDP address/port.