Ensure the flush timer is started if we never had any full packets to send.
The flush timer ensures that if there is partial data left in the send
buffer for framed UDP channels, that this is sent after the flush timeout.
This was done correctly if we had previously sent full packets, but if there
never was a full packet the timer wouldn't be started, and the small frames
wouldn't be sent unless the application would send more data on the channel.