Wake up the MeshLink thread if framed channel data is pending to be flushed.
When sending data on framed UDP channels, if there is a partial frame in
the send buffer waiting to be flushed after the flush timer expires, this
data was added by the application thread. The MeshLink thread does not know
that a timer was updated, and might use an old timeout value and not
respond in time. So if we detect that this might happen, we signal the
MeshLink thread so it can calculate a new timeout and call select() again.