Allow incoming connections to specify a real address in the ID request.
If there is some kind of proxy that is forwarding meta-connections,
then MeshLink will see incoming connections with the address of the proxy
instead of the real address of the peer. This can be problematic for other
peers that want to form meta-connections. So allow an address and port to be
appended to ID requests, that will cause MeshLink to announce that address
to other nodes instead of the one from the incoming meta-connection.