Try to recover from select() returning an error.
If we get an error back from select(), it could be because one of the
filedescriptors is in a bad state. This shouldn't occur, but if it does,
just call all the registered I/O callbacks to have them check for errors.
Only quit the event loop if the error persists.

Also fix some I/O callbacks that didn't expect to be called when the
filedescriptor isn't signalled.