Ignore ADD_EDGE messages we know are outdated.
A node reconnecting after a network error might send outdated ADD_EDGE
messages. If we see this for an edge we own, we send back a correction.
However, if we don't own the edge but we are one of the edges endpoints,
we ignore it and let the other side, which is the owner, take care of
sending a correction.

This change avoids us from incorrectly learning an old session ID for a
node, which might cause channels to that node to be closed prematurely.