Use a key/value store with configurable storage callbacks.
For greater flexibility with how configuration files are stored, we
assume a key/value store, and allow the application to provide callbacks
for storing and loading values, and listing all the keys. By default, we
use the filesystem as a backing store, with one file per key in a flat
directory.